requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# AI & LLM
anthropic==0.34.0
langchain==0.3.0
langchain-anthropic==0.2.0
langchain-openai==0.2.0

# Vector store
pgvector==0.3.3
qdrant-client==1.11.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Cache / session
redis==5.1.0

# HTTP client (untuk ERP & WAHA)
httpx==0.27.2

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0        # retry logic
structlog==24.4.0      # structured logging
